+++ evalml/pipelines/classification/catboost.py
@@ -1,6 +1,5 @@
 from skopt.space import Integer, Real
 
-from evalml.model_family import ModelFamily
 from evalml.pipelines import PipelineBase
 
 
@@ -12,8 +11,6 @@ class CatBoostClassificationPipeline(PipelineBase):
     For more information, check out https://catboost.ai/
     Note: impute_strategy must support both string and numeric data
     """
-    name = "CatBoost Classifier w/ Simple Imputer"
-    model_family = ModelFamily.CATBOOST
     component_graph = ['Simple Imputer', 'CatBoost Classifier']
     problem_types = ['binary', 'multiclass']
     hyperparameters = {