+++ evalml/pipelines/classification/logistic_regression.py
@@ -5,7 +5,6 @@ from evalml.pipelines import PipelineBase
 
 class LogisticRegressionPipeline(PipelineBase):
     """Logistic Regression Pipeline for both binary and multiclass classification"""
-    name = "Logistic Regression Classifier w/ One Hot Encoder + Simple Imputer + Standard Scaler"
     component_graph = ['One Hot Encoder', 'Simple Imputer', 'Standard Scaler', 'Logistic Regression Classifier']
     problem_types = ['binary', 'multiclass']