+++ evalml/pipelines/pipeline_base.py
@@ -1,4 +1,5 @@
 import copy
+import re
 from abc import ABC, abstractmethod
 from collections import OrderedDict
 
@@ -8,6 +9,7 @@ from sklearn.model_selection import train_test_split
 from .components import Estimator, handle_component
 from .graphs import make_feature_importance_graph, make_pipeline_graph
 
+from evalml.exceptions import IllFormattedClassNameError
 from evalml.objectives import get_objective
 from evalml.problem_types import handle_problem_types
 from evalml.utils import Logger, classproperty
@@ -54,22 +56,43 @@ class PipelineBase(ABC):
         if self.estimator is None:
             raise ValueError("A pipeline must have an Estimator as the last component in component_graph.")
 
-        self.name = self._generate_name()
         self._validate_problem_types(self.problem_types)
 
-    def _generate_name(self):
-        "Generates name from components in self.component_graph"
-        if self.estimator is not None:
-            name = "{}".format(self.estimator.name)
-        else:
-            name = "Pipeline"
-        for index, component in enumerate(self.component_graph[:-1]):
-            if index == 0:
-                name += " w/ {}".format(component.name)
-            else:
-                name += " + {}".format(component.name)
-
+    @classproperty
+    def name(cls):
+        """Returns a name describing the pipeline.
+        By default, this will take the class name and add a space between each capitalized word. If the pipeline has a _name attribute, this will be returned instead.
+        """
+        try:
+            name = cls._name
+        except AttributeError:
+            rex = re.compile(r'(?<=[a-z])(?=[A-Z])')
+            name = rex.sub(' ', cls.__name__)
+            if name == cls.__name__:
+                raise IllFormattedClassNameError("Pipeline Class {} needs to follow pascall case standards or `_name` must be defined.".format(cls.__name__))
         return name
+
+    @classproperty
+    def summary(cls):
+        """Returns a short summary of the pipeline structure, describing the list of components used.
+        Example: Logistic Regression Classifier w/ Simple Imputer + One Hot Encoder
+        """
+        def _generate_summary(component_graph):
+            component_graph[-1] = handle_component(component_graph[-1])
+            estimator = component_graph[-1] if isinstance(component_graph[-1], Estimator) else None
+            if estimator is not None:
+                summary = "{}".format(estimator.name)
+            else:
+                summary = "Pipeline"
+            for index, component in enumerate(component_graph[:-1]):
+                component = handle_component(component)
+                if index == 0:
+                    summary += " w/ {}".format(component.name)
+                else:
+                    summary += " + {}".format(component.name)
+            return summary
+
+        return _generate_summary(cls.component_graph)
 
     def _validate_problem_types(self, problem_types):
         """Validates provided `problem_types` against the estimator in `self.component_graph`