+++ evalml/pipelines/regression/linear_regression.py
@@ -3,7 +3,6 @@ from evalml.pipelines import PipelineBase
 
 class LinearRegressionPipeline(PipelineBase):
     """Linear Regression Pipeline for regression problems"""
-    name = "Linear Regressor w/ One Hot Encoder + Simple Imputer + Standard Scaler"
     component_graph = ['One Hot Encoder', 'Simple Imputer', 'Standard Scaler', 'Linear Regressor']
     problem_types = ['regression']